requirements.txt
scikit-learn==1.5.1
numpy==1.26.4
scipy==1.13.1
jsonschema==4.23.0
pytest==8.3.2
joblib==1.4.2
sentence-transformers==3.3.1
faiss-cpu==1.8.0
pyyaml==6.0.2
jinja2>=3.1.0
google-generativeai==0.8.3
openai==1.51.0
python-dotenv==1.0.1


